check to see if a template was stashed before going to the default

+++ sim/memlcd/dut.py
@@ -147,7 +147,10 @@ def run_sim(ci=False):
     
     # initialize with a default waveform that contains the most basic execution tracing
     if os.path.isfile('run/top_tb_sim.wcfg') != True:
-        os.system('cp ../../sim_support/top_tb_sim.wcfg run/')
+        if os.path.isfile('top_tb_sim.wcfg'):
+            os.system('cp top_tb_sim.wcfg run/')
+        else:
+            os.system('cp ../../sim_support/top_tb_sim.wcfg run/')
 
     # load up simulator dependencies
     os.system("cd run && cp gateware/*.init .")
